requirements.txt
fastapi==0.111.0
pymongo==4.7.1
python-dateutil==2.9.0.post0
